container backup: rename ambiguous variables and files
